Change the algorithm to order the commit.
In some situations, the representation was not correct due to a bad
commit order.
Now there is a check on parents and brother to guarantee to place the
commit on the right place.
However, it is sometimes impossible to represent the graph correctly
when there were commits manipulations or move.
Typically when a commits comes from a merge of 2 commits of the same
branch. In this case, as suffix is added "!!!" to notify the user that
the representation is not 100% correct.